Move around some utils
