+++ src/triage/component/architect/utils.py
@@ -4,10 +4,6 @@ import operator
 from triage.util.structs import FeatureNameList
 
 
-def str_in_sql(values):
-    return ",".join(map(lambda x: "'{}'".format(x), values))
-
-
 def feature_list(feature_dictionary):
     """Convert a feature dictionary to a sorted list
