--- src/triage/component/audition/utils.py
@@ -1,6 +1,0 @@
-def make_list(a):
-    return [a] if not isinstance(a, list) else a
-
-
-def str_in_sql(values):
-    return ",".join(map(lambda x: "'{}'".format(x), values))
