+++ src/triage/util/db.py
@@ -1,11 +1,18 @@
 # coding: utf-8
 
+import csv
 import sqlalchemy
 import wrapt
 from contextlib import contextmanager
 from sqlalchemy.orm import Session
 from sqlalchemy.engine.url import make_url
+from retrying import retry
+from functools import partial
+from itertools import chain
 
+import postgres_copy
+
+from ohio import PipeTextIO
 
 class SerializableDbEngine(wrapt.ObjectProxy):
     """A sqlalchemy engine that can be serialized across process boundaries.
@@ -58,3 +65,51 @@ def get_for_update(db_engine, orm_class, primary_key):
         obj = session.query(orm_class).get(primary_key)
         yield obj
         session.merge(obj)
+
+
+def retry_if_db_error(exception):
+    return isinstance(exception, sqlalchemy.exc.OperationalError)
+
+
+DEFAULT_RETRY_KWARGS = {
+    "retry_on_exception": retry_if_db_error,
+    "wait_exponential_multiplier": 1000,  # wait 2^x*1000ms between each retry
+    "stop_max_attempt_number": 14,
+    # with this configuration, last wait will be ~2 hours
+    # for a total of ~4.5 hours waiting
+}
+
+
+db_retry = retry(**DEFAULT_RETRY_KWARGS)
+
+
+def _write_csv(file_like, db_objects, type_of_object):
+    writer = csv.writer(file_like, quoting=csv.QUOTE_MINIMAL, lineterminator='\n')
+    for db_object in db_objects:
+        if type(db_object) != type_of_object:
+            raise TypeError("Cannot copy collection of objects to db as they are not all "
+                            f"of the same type. First object was {type_of_object} "
+                            f"and later encountered a {type(db_object)}")
+        writer.writerow(
+            [getattr(db_object, col.name) for col in db_object.__table__.columns]
+        )
+
+
+@db_retry
+def save_db_objects(db_engine, db_objects):
+    """Saves a collection of SQLAlchemy model objects to the database using a COPY command
+
+    Args:
+        db_engine (sqlalchemy.engine)
+        db_objects (iterable) SQLAlchemy model objects, corresponding to a valid table
+    """
+    db_objects = iter(db_objects)
+    first_object = next(db_objects)
+    type_of_object = type(first_object)
+
+    with PipeTextIO(partial(
+            _write_csv,
+            db_objects=chain((first_object,), db_objects),
+            type_of_object=type_of_object
+    )) as pipe:
+        postgres_copy.copy_from(pipe, type_of_object, db_engine, format="csv")
