+++ src/triage/util/hash.py
@@ -1,0 +1,14 @@
+import datetime
+import hashlib
+import json
+
+
+def filename_friendly_hash(inputs):
+    def dt_handler(x):
+        if isinstance(x, datetime.datetime) or isinstance(x, datetime.date):
+            return x.isoformat()
+        raise TypeError("Unknown type")
+
+    return hashlib.md5(
+        json.dumps(inputs, default=dt_handler, sort_keys=True).encode("utf-8")
+    ).hexdigest()
