+++ src/triage/util/iteration.py
@@ -1,0 +1,27 @@
+class Batch:
+    # modified from
+    # http://codereview.stackexchange.com/questions/118883/split-up-an-iterable-into-batches
+    def __init__(self, iterable, limit=None):
+        self.iterator = iter(iterable)
+        self.limit = limit
+        try:
+            self.current = next(self.iterator)
+        except StopIteration:
+            self.on_going = False
+        else:
+            self.on_going = True
+
+    def group(self):
+        yield self.current
+        # start enumerate at 1 because we already yielded the last saved item
+        for num, item in enumerate(self.iterator, 1):
+            self.current = item
+            if num == self.limit:
+                break
+            yield item
+        else:
+            self.on_going = False
+
+    def __iter__(self):
+        while self.on_going:
+            yield self.group()
